# JIT compilation, first cut, squashed