--- a/console/runner.ts
+++ b/console/runner.ts
@@ -46,6 +46,7 @@ javaCLI(process.argv.slice(2), {
   nativeClasspath: [path.resolve(__dirname, '../src/natives')],
   launcherName: process.argv[0] + " " + path.relative(process.cwd(), process.argv[1]),
   intMode: false,
+  dumpJITStats: false,
   tmpDir: os.tmpdir()
 }, doneCb, function(jvm: JVM): void {
   jvmState = jvm;