--- a/src/java_cli.ts
+++ b/src/java_cli.ts
@@ -42,6 +42,9 @@ let parser = new OptionParser({
     'int': {
       desc: 'interpreted mode execution only'
     },
+    'dump-JIT-stats': {
+      desc: 'dump JIT statistics'
+    },
     log: {
       desc: 'log level, [0-10]|vtrace|trace|debug|error',
       type: ParseType.NORMAL_VALUE_SYNTAX
@@ -116,6 +119,7 @@ function java(args: string[], opts: JVMCLIOptions,
   let logOption = nonStandard.stringOption('log', 'ERROR');
 
   opts.intMode = nonStandard.flag('int', false);
+  opts.dumpJITStats = nonStandard.flag('dump-JIT-stats', false);
 
   if (/^[0-9]+$/.test(logOption)) {
     logging.log_level = parseInt(logOption, 10);