--- a/src/java_cli.ts
+++ b/src/java_cli.ts
@@ -82,6 +82,9 @@ let parser = new OptionParser({
       type: ParseType.COLON_VALUE_SYNTAX,
       optDesc: ':<directories and zip/jar files separated by :>',
       desc: 'set search path for bootstrap classes and resources'
+    },
+    'PrintCompilation': {
+      desc: 'Print JIT compilation details'
     }
   }
 });
@@ -207,6 +210,8 @@ function java(args: string[], opts: JVMCLIOptions,
       launchJvm(standard, opts, jvmState, doneCb, jvmStarted);
     }
   });
+
+  jvmState.setPrintJITCompilation(nonStandard.flag('PrintCompilation', false));
 
   let vtraceMethods = nonStandard.stringOption('vtrace-methods', null);
   if (vtraceMethods) {