… tasks/test/harness_common.ts