@jvilk 2016-05-14
I've started looking at this now! My first comments are purely organizational in nature; I have not thought deeply about the JIT mechanisms yet. (That is coming next!)

I am slightly worried about bifurcating opcode logic into multiple files for maintenance reasons. I wonder if it would be cleaner to have a `class` for each opcode type with `static`-only methods and fields? That would collect the following information for each opcode into one place:

* The interpreter function
* The function that emits JavaScript for the JIT.
* Useful JIT properties (e.g. whether or not it is a JUMP instruction)
  * Then, you could change the code in the JIT that hardcodes particular opcodes to simply check if the opcode has a certain property (e.g. `if (opcodetype.jumps())` instead of `if (opcode === INVOKESTATIC || (etc))`
(etc.)

Each `class` itself would implement a consistent interface, so the result would be strongly typed as well.

The interpreter would gather all of the static interpreter functions into its lookup table like it does now, so it would not change any of that logic; it would just change the source of the interpreter functions.

Do you think that would result in cleaner, easier-to-maintain code?