@jvilk 2016-05-14
I'm going to make this a meta-comment with a checklist, to be periodically updated as I notice things.

* [ ] (If agreed) Consolidate opcode logic into `class`es with `static` methods and fields.
* [ ] Expose `-Xint` option to switch off the JIT and confine execution to the interpreter.