@jvilk 2016-05-14
OK, I've looked this over. Overall, this is great work, and I am excited to tinker with it once it is merged! :)

Current blockers to merging:

* The heuristic used to determine if something should be JITed is expensive and should be cheaper
    * Using some combination of number of executions of the function/basic block and function size seems like a good idea. Maybe we should conduct a literature search for common JIT compiler heuristics to experiment with.
* There should be an API and command-line flag to disable JITting (`-Xint`)
* The JIT should produce a single function that supplants the method's `run` function to avoid the shenanigans it currently does in `getOp`. :)

Do these changes sound reasonable?