@hrj 2016-05-16
> Do you think that would result in cleaner, easier-to-maintain code?

I am not sure.. in some ways it is nicer to have separate files for interpreted v/s JIT code. Atleast while developing, I found it easier to have separate files, because each file had its own pattern and searching by opcode was easy.