@hrj 2016-05-16
> Useful JIT properties (e.g. whether or not it is a JUMP instruction)
> 
>     Then, you could change the code in the JIT that hardcodes particular opcodes to simply check if the opcode has a certain property (e.g. if (opcodetype.jumps()) instead of if (opcode === INVOKESTATIC || (etc)) (EDIT: I guess you have .hasBranches! But it would formalize all of that into one place.) (etc.)

This is already in place. There is an interface called `JitInfo` defined in `jit.ts` that captures this info.

The `INVOKE_*` methods are special cases because their information is more dynamic than the other opcodes. Specifically, the number of operands popped by an `invoke` opcode depends on the method being called, unlike most other opcodes which have a fixed number of "pops".