@jvilk 2016-05-16
> This is already in place. There is an interface called JitInfo defined in jit.ts that captures this info.

I guess my frustration is that it's not easy to extend the current setup with other, more dynamic info without further hacks. For example, for opcodes that make a NULL check, expose a function that determines the symbolic value that it checks. Then, under certain conditions, you can remove redundant NULL checks.

Anyway, any such changes could come post-merge, if it makes sense. I'll remove it from the blocking list.