@hrj 2016-05-16
> I guess my frustration is that it's not easy to extend the current setup with other, more dynamic info without further hacks. For example, for opcodes that make a NULL check, expose a function that determines the symbolic value that it checks. Then, under certain conditions, you can remove redundant NULL checks.

For that specific example, is it possible that javascript VM can figure it out and eliminate the check? I have had only a brief glance at the debug log of V8 and it does seem like they have some aggressive optimisations.

> 
> Anyway, any such changes could come post-merge, if it makes sense. I'll remove it from the blocking list.

Yes, I think we could do it post-merge. If the approach of this PR seems good, we could use it as a baseline for performance and incrementally add many more features to the JIT.