@jvilk 2016-05-16
> is it possible that javascript VM can figure it out and eliminate the check?

I have my doubts, since the return value from the `isNull` method is used in a conditional. But there is a way to check this -- you can actually use [IR viewers for V8](http://mrale.ph/irhydra/1/) to take a look at the resulting IR to see if it removes branches.

I'm pretty sure it will not remove it, though.

>Yes, I think we could do it post-merge. If the approach of this PR seems good, we could use it as a baseline for performance and incrementally add many more features to the JIT.

Sounds good to me. I updated the checklist; once those items are addressed, I will merge this in!