@jvilk 2016-05-18
I'd like to get rid of those conditionals somehow (`if (offset < 0)`), but it's OK for now. :)

> I was not able to add -XX:+PrintCompilation option. I have committed my attempt in d0ea20a but the option parser is complaining about something I don't understand. I could do this post-merge.

Ahh, OK. I wrote that parser. Could you add it as `-XPrintCompilation` for now, and I'll fix the parser post-merge?

Thanks for being patient! We're so close to merging.