@hrj 2016-05-18
> Could you add it as -XPrintCompilation for now, and I'll fix the parser post-merge?

I have started working on this, but I worry that this feature might affect performance adversely, so I wanted to clarify. The flag is stored in the `JVM`, which is accessible via the thread object. So to access the flag in the JIT compiler methods, we need to pass the `thread` object along the call chain. Is that correct? Or is there a better way to access such flags from a `Method`'s context?

> 
> Thanks for being patient! We're so close to merging.

No worries; this PR would have a big impact so it's good to take our time with it.