@hrj 2016-05-18
Also, sorry I haven't been able to address all points raised by you. There are many ideas and options to explore here and the need to benchmark and discuss alternatives. It might be best to create separate issues for each topic after this gets merged.